Keep upstream e2e test separate
